requirements.txt
PySide6>=6.6
pynput>=1.7.6
pywin32>=306
mss>=9.0
Pillow>=10.0
